# Conflicts:
#	src/biopsykit/signals/ecg/ecg.py

+++ tests/test_questionnaires/test_questionnaires_utils.py
@@ -6,6 +6,9 @@ from unittest import TestCase
 import numpy as np
 import pandas as pd
 import pytest
+from numpy.testing import assert_array_equal
+from pandas._testing import assert_frame_equal, assert_series_equal
+
 from biopsykit.questionnaires.utils import (
     bin_scale,
     compute_scores,
@@ -19,8 +22,6 @@ from biopsykit.questionnaires.utils import (
     zero_pad_columns,
 )
 from biopsykit.utils.exceptions import ValidationError, ValueRangeError
-from numpy.testing import assert_array_equal
-from pandas._testing import assert_frame_equal, assert_series_equal
 
 TEST_FILE_PATH = Path(__file__).parent.joinpath("../test_data/questionnaires")
 
@@ -981,9 +982,11 @@ class TestQuestionnairesUtils:
     )
     def test_bin_scale(self, data, bins, cols, first_min, last_max, inplace, expected_in, expected_out):
         out = bin_scale(data=data, bins=bins, cols=cols, first_min=first_min, last_max=last_max, inplace=inplace)
-
+        data = data.astype("Int64")
+        expected_in = expected_in.astype("Int64")
         assert_frame_equal(data, expected_in)
         if expected_out is not None:
+            expected_out = expected_out.astype("Int64")
             assert_frame_equal(out, expected_out)
 
     @pytest.mark.parametrize(
@@ -1046,6 +1049,7 @@ class TestQuestionnairesUtils:
 
         assert_series_equal(data, expected_in)
         if expected_out is not None:
+            expected_out = expected_out.astype("Int64")
             assert_series_equal(out, expected_out)
 
     def test_wide_to_long_warning(self):
